requirements.txt
Django==1.11.28
djangorestframework==3.9.1
drf-yasg==1.17.0
pytz==2016.10
